Add a very basic report format support.
Basically, we have a Scanner.ScanReport() object that has
a dict of SshJob results (which include the rho_cmds results).

The report just walks over those, generating a report based on
simple python string formatting.

This may be enough, or maybe we need template support, or smarter
report classes, but this is a start.

make the default report format a csv format that I just made up