add support for scanning profiles and using their auth or a auth specified on the cli
move _create_range to base class since we reuse twice(guess we should
make an intermiddle class for the subclasses that use it maybe?)

If we specify auth info or profile info on the "scan" cli, we create
auth and profile classes of "clioption" type so we can reuse the
code that expects profile/auth to be in the config dicts. We don't
save these options.