Try to handle the cases of hostnames being in the ranges a little better.
By default, we now expect ' - ' as the seperator in the profile "ranges"
paramater. This makes it a little easier to handle hostnames here.
WARNING: this may break existing configs.

Rename _check_for_hostname to _find_ip and change it's behaviour.
Basically, it tries to get a meaningful ip from what you give it. Note
however, that the stuff to determine if a string is range/cidr/etc
happens _before_ this is called.