Try to validate ip ranges given on the command line.
Add _validate_ranges to base command class. It takes a ip range list
and returns true if the ranges look valid, false otherwise.

in rho_ips use socket.inet_aton to see if we have a valid looking
ip address. Set self.valid = False on the RhoIpRange object if
some of it's inputs are bogus.

bugs: * profile add --range:  the range string should be validated for
	format at  profile creation (fixed)