refactor the output for Profiles.
Printing profiles using a table caused problems when there were long
lists of ranges, ports, or auths. We tried creating a tabular output
but it caused a false association between ports and ranges.

i.e.

name           ranges                      ports   auths
profilename    127.0.0.1                   22      foo
               192.168.1.1 - 192.168.1.23  2222    auth1
                                           5422

This seems to imply that 127.0.0.1 is associated with port 22
and auth foo. This is not the case as all ports and all auths apply
to each of the ranges.

Resulting output is more list like in nature:

bin/rho profile list
Config Encryption Password:

name: this is a really long name
    auths:
        foo
        this is a long ass name
    ports:
        22
    ranges:
        127.0.0.1

name: sigh
    auths:
        foo
    ports:
        22
        2222
        5422
    ranges:
        192.168.1.1 - 192.168.1.24