fix auth (add|edit) man page entries.
The man page erroneously stated you could pass in a password option to auth add.
This is incorrect as that is an invalid option for that sub command.

The man page erroneously stated you could pass in a password to the auth edit
command. That is incorrect, while the --password IS a valid option, it does not
take any parameters. The user will be prompted for the password to avoid leaking
it via the CLI.