1395080 - exit gracefully on missing report
if report doesn't exist exit gracefully. If it is pack-scan, the
default report, the user didn't run initconfig. Instead of
exiting we simply add it to the in-memory config before running
the scan. While it seems like a good idea to write the report to
the config, we modify the config with other temporary runtime
options which would pollute the config if written. For now, we
are opting to simply add the report in-memory.